+++ tests/tests_elements/tests_cycle_decomposition/test_cases.py
@@ -47,6 +47,11 @@ TEST_CYCLE_TYPE = [
     (CycleDecomposition(Cycle(1, 3, 2), Cycle(4)), (1, 3)),
     (CycleDecomposition(Cycle(1, 2), Cycle(3, 4)), (2, 2)),
 ]
+TEST_DEGREE = [
+    (CycleDecomposition(Cycle(1)), 1),
+    (CycleDecomposition(Cycle(1), Cycle(2, 3)), 3),
+    (CycleDecomposition(Cycle(1), Cycle(2), Cycle(3)), 3),
+]
 TEST_DESCENTS = [
     (CycleDecomposition(Cycle(1, 2, 3)), [2]),
     (CycleDecomposition(Cycle(1), Cycle(2), Cycle(3)), []),